+++ modules/member/controllers/BrevetawardController.php
@@ -8,6 +8,7 @@ use app\modules\member\searchs\BrevetAwardSerch;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
+use yii\web\HttpException;
 
 /**
  * BrevetawardController implements the CRUD actions for BrevetAwardModel model.
@@ -33,13 +34,17 @@ class BrevetawardController extends Controller
      */
     public function actionIndex()
     {
-        $searchModel = new BrevetAwardSerch;
-        $dataProvider = $searchModel->search(Yii::$app->request->getQueryParams());
+        if (Yii::$app->user->can('brevetawardindex')) {
+            $searchModel = new BrevetAwardSerch;
+            $dataProvider = $searchModel->search(Yii::$app->request->getQueryParams());
 
-        return $this->render('index', [
-                    'dataProvider' => $dataProvider,
-                    'searchModel' => $searchModel,
-        ]);
+            return $this->render('index', [
+                        'dataProvider' => $dataProvider,
+                        'searchModel' => $searchModel,
+            ]);
+        } else {
+            throw new HttpException(403, 'You are not allowed to access this page', 0);
+        }
     }
 
     /**
@@ -49,9 +54,13 @@ class BrevetawardController extends Controller
      */
     public function actionView($id)
     {
-        return $this->render('view', [
-                    'model' => $this->findModel($id),
-        ]);
+        if (Yii::$app->user->can('brevetawardview')) {
+            return $this->render('view', [
+                        'model' => $this->findModel($id),
+            ]);
+        } else {
+            throw new HttpException(403, 'You are not allowed to access this page', 0);
+        }
     }
 
     /**
@@ -61,16 +70,20 @@ class BrevetawardController extends Controller
      */
     public function actionCreate()
     {
-        $model = new BrevetAwardModel;
-        $model->setAttribute('term_id', MEMBER_BREVET);
-        $model->setAttribute('create_et', date("Y-m-d H:i:s"));
-        $model->setAttribute('update_et', date("Y-m-d H:i:s"));
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'action' => 'member-brevet-view', 'id' => $model->id]);
+        if (Yii::$app->user->can('brevetawardcreate')) {
+            $model = new BrevetAwardModel;
+            $model->setAttribute('term_id', MEMBER_BREVET);
+            $model->setAttribute('create_et', date("Y-m-d H:i:s"));
+            $model->setAttribute('update_et', date("Y-m-d H:i:s"));
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'action' => 'member-brevet-view', 'id' => $model->id]);
+            } else {
+                return $this->render('create', [
+                            'model' => $model,
+                ]);
+            }
         } else {
-            return $this->render('create', [
-                        'model' => $model,
-            ]);
+            throw new HttpException(403, 'You are not allowed to access this page', 0);
         }
     }
 
@@ -82,24 +95,32 @@ class BrevetawardController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
-        $model->setAttribute('update_et', date("Y-m-d H:i:s"));
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'action' => 'member-brevet-view', 'id' => $model->id]);
+        if (Yii::$app->user->can('brevetawardupdate')) {
+            $model = $this->findModel($id);
+            $model->setAttribute('update_et', date("Y-m-d H:i:s"));
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'action' => 'member-brevet-view', 'id' => $model->id]);
+            } else {
+                return $this->render('update', [
+                            'model' => $model,
+                ]);
+            }
         } else {
-            return $this->render('update', [
-                        'model' => $model,
-            ]);
+            throw new HttpException(403, 'You are not allowed to access this page', 0);
         }
     }
 
     public function actionBulk()
     {
-        if (Yii::$app->request->post() && (Yii::$app->request->post('bulk_action1') == 'delete' || Yii::$app->request->post('bulk_action2') == 'delete')) {
-            $this->deleteAll(Yii::$app->request->post('selection'));
-            return $this->redirect(['index','action'=>'member-brevet-list']);
+        if (Yii::$app->user->can('brevetawardbulk')) {
+            if (Yii::$app->request->post() && (Yii::$app->request->post('bulk_action1') == 'delete' || Yii::$app->request->post('bulk_action2') == 'delete')) {
+                $this->deleteAll(Yii::$app->request->post('selection'));
+                return $this->redirect(['index', 'action' => 'member-brevet-list']);
+            } else {
+                return $this->redirect(['index', 'action' => 'member-brevet-list']);
+            }
         } else {
-            return $this->redirect(['index','action'=>'member-brevet-list']);
+            throw new HttpException(403, 'You are not allowed to access this page', 0);
         }
     }
 
@@ -111,9 +132,13 @@ class BrevetawardController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+        if (Yii::$app->user->can('brevetawarddelete')) {
+            $this->findModel($id)->delete();
 
-        return $this->redirect(['index','action'=>'member-brevet-list']);
+            return $this->redirect(['index', 'action' => 'member-brevet-list']);
+        } else {
+            throw new HttpException(403, 'You are not allowed to access this page', 0);
+        }
     }
 
     /**
@@ -127,7 +152,7 @@ class BrevetawardController extends Controller
                 $this->findModel($id)->delete();
             }
         } else {
-            return $this->redirect(['index','action'=>'member-brevet-list']);
+            return $this->redirect(['index', 'action' => 'member-brevet-list']);
         }
     }
