+++ modules/member/controllers/CapasController.php
@@ -1,6 +1,7 @@
 <?php
 
 namespace app\modules\member\controllers;
+
 define("DS", DIRECTORY_SEPARATOR);
 
 use Yii;
@@ -9,41 +10,30 @@ use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
 use app\modules\member\models\LifeSkillModel;
-use app\modules\filemanager\librarys\Image;
+use yii\web\HttpException;
 use app\modules\member\models\CapasModel;
 use app\modules\member\searchs\CapasSerch;
-
+use app\modules\member\helpers\PhotoHelper;
 
 /**
  * CapasController implements the CRUD actions for Member model.
  */
 class CapasController extends Controller
 {
+
     public $langskill;
     public $lifeskill;
     public $otherlifeskill;
     public $brevetaward;
     public $other_skill;
     public $otherlifeskillid;
-
-    public $path;
-    public $original;
-    public $certificate;
-    public $identitycard;
-    public $frontphoto;
-    public $sidephoto;
+    public $photo;
 
     public function init()
     {
-        $this->path = Yii::$app->getBasePath() . DS . 'web' . DS . 'resources' . DS . 'images' . DS;
-        $this->original = $this->path . 'original' . DS;
-        $this->frontphoto = $this->path . 'member' . DS . 'frontphoto' . DS;
-        $this->sidephoto = $this->path . 'member' . DS . 'sidephoto' . DS;
-        $this->certificate = $this->path . 'member' . DS . 'certificate' . DS;
-        $this->identitycard = $this->path . 'member' . DS . 'identitycard' . DS;
+        $this->photo = new PhotoHelper;
         parent::init();
     }
-
 
     public function behaviors()
     {
@@ -57,25 +47,28 @@ class CapasController extends Controller
         ];
     }
 
-
     /**
      * Lists all CapasModel models.
      * @return mixed
      */
     public function actionIndex()
     {
-        $searchModel = new CapasSerch;
-        $dataProvider = $searchModel->search(Yii::$app->request->getQueryParams());
-        if ((Yii::$app->request->get('bulk_action1') == 'delete' || Yii::$app->request->get('bulk_action2') == 'delete')) {
-            $this->deleteAll(Yii::$app->request->get('selection'));
+        if (Yii::$app->user->can('capasindex')) {
+            $searchModel = new CapasSerch;
+            $dataProvider = $searchModel->search(Yii::$app->request->getQueryParams());
+            if ((Yii::$app->request->get('bulk_action1') == 'delete' || Yii::$app->request->get('bulk_action2') == 'delete')) {
+                $this->deleteAll(Yii::$app->request->get('selection'));
+            }
+            if ((Yii::$app->request->get('bulk_action1') == 'trash' || Yii::$app->request->get('bulk_action2') == 'trash')) {
+                $this->trashAll(Yii::$app->request->get('selection'));
+            }
+            return $this->render('index', [
+                        'dataProvider' => $dataProvider,
+                        'searchModel' => $searchModel,
+            ]);
+        } else {
+            throw new HttpException(403, 'You are not allowed to access this page', 0);
         }
-        if ((Yii::$app->request->get('bulk_action1') == 'trash' || Yii::$app->request->get('bulk_action2') == 'trash')) {
-            $this->trashAll(Yii::$app->request->get('selection'));
-        }
-        return $this->render('index', [
-            'dataProvider' => $dataProvider,
-            'searchModel' => $searchModel,
-        ]);
     }
 
     /**
@@ -85,9 +78,13 @@ class CapasController extends Controller
      */
     public function actionView($id)
     {
-        return $this->render('view', [
-            'model' => $this->findModel($id),
-        ]);
+        if (Yii::$app->user->can('capasview')) {
+            return $this->render('view', [
+                        'model' => $this->findModel($id),
+            ]);
+        } else {
+            throw new HttpException(403, 'You are not allowed to access this page', 0);
+        }
     }
 
     /**
@@ -97,25 +94,25 @@ class CapasController extends Controller
      */
     public function actionCreate()
     {
-        $model = new CapasModel;
-        $model->setAttribute('create_et', date("Y-m-d H:i:s"));
-        $model->setAttribute('update_et', date("Y-m-d H:i:s"));
-        $model->setAttribute('type_member', MEMBER_TYPE_CAPAS);
-        $model->setAttribute('user_id', Yii::$app->user->identity->getId());
+        if (Yii::$app->user->can('capascreate')) {
+            $model = new CapasModel;
+            $model->setAttribute('create_et', date("Y-m-d H:i:s"));
+            $model->setAttribute('update_et', date("Y-m-d H:i:s"));
+            $model->setAttribute('type_member', MEMBER_TYPE_CAPAS);
+            $model->setAttribute('user_id', Yii::$app->user->identity->getId());
 
-        if ($model->load(Yii::$app->request->post())) {
+            $param = Yii::$app->request->post('CapasModel');
+            $this->langskill = $param['language_skill'];
+            $this->lifeskill = $param['life_skill'];
+            $this->otherlifeskill = $param['otherlifeskill'];
+            $this->brevetaward = $param['brevet_award'];
 
-            $form = Yii::$app->request->post('CapasModel');
-            $this->langskill = $form['language_skill'];
-            $this->lifeskill = $form['life_skill'];
-            $this->otherlifeskill = $form['otherlifeskill'];
-            $this->brevetaward = $form['brevet_award'];
+            $this->photo->copyProntPhoto($param['front_photo']);
+            $this->photo->copyProntPhotoThumb($param['front_photo']);
+            $this->photo->copySidePhoto($param['side_photo']);
+            $this->photo->copyCertificatePhoto($param['certificate_of_organization']);
+            $this->photo->copyIdentityCardPhoto($param['identity_card']);
 
-            $this->copyProntPhoto($form['front_photo']);
-            $this->copyProntPhotoThumb($form['front_photo']);
-            $this->copySidePhoto($form['side_photo']);
-            $this->copyCertificatePhoto($form['certificate_of_organization']);
-            $this->copyIdentityCardPhoto($form['identity_card']);
 
             if (null != $this->otherlifeskill) {
                 $this->saveOtherLifeSkill($this->otherlifeskill);
@@ -142,7 +139,9 @@ class CapasController extends Controller
                 $data_brevet = $model->getAllBrevetNameById($this->brevetaward);
                 $model->setAttribute('brevetaward', implode(', ', $data_brevet));
             }
-            if ($model->save()) {
+
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+
                 if (null != $this->lifeskill) {
                     $model->saveTaxRelation($this->lifeskill, $model->id);
                 }
@@ -156,15 +155,13 @@ class CapasController extends Controller
                 }
                 return $this->redirect(['view', 'action' => 'member-capas-view', 'id' => $model->id]);
             } else {
+
                 return $this->render('create', [
-                    'model' => $model,
+                            'model' => $model,
                 ]);
             }
         } else {
-
-            return $this->render('create', [
-                'model' => $model,
-            ]);
+            throw new HttpException(403, 'You are not allowed to access this page', 0);
         }
     }
 
@@ -176,21 +173,17 @@ class CapasController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
-        $model->setAttribute('update_et', date("Y-m-d H:i:s"));
-        $model->setAttribute('user_id', Yii::$app->user->identity->getId());
-        if ($model->load(Yii::$app->request->post())) {
-            $form = Yii::$app->request->post('CapasModel');
-            $this->langskill = $form['language_skill'];
-            $this->lifeskill = $form['life_skill'];
-            $this->otherlifeskill = $form['otherlifeskill'];
-            $this->brevetaward = $form['brevet_award'];
+        if (Yii::$app->user->can('capasupdate')) {
+            $model = $this->findModel($id);
+            $model->setAttribute('update_et', date("Y-m-d H:i:s"));
+            $model->setAttribute('user_id', Yii::$app->user->identity->getId());
+            $param = Yii::$app->request->post('CapasModel');
 
-            $this->copyProntPhoto($form['front_photo']);
-            $this->copyProntPhotoThumb($form['front_photo']);
-            $this->copySidePhoto($form['side_photo']);
-            $this->copyCertificatePhoto($form['certificate_of_organization']);
-            $this->copyIdentityCardPhoto($form['identity_card']);
+            //skill
+            $this->langskill = $param['language_skill'];
+            $this->lifeskill = $param['life_skill'];
+            $this->otherlifeskill = $param['otherlifeskill'];
+            $this->brevetaward = $param['brevet_award'];
 
             if (null != $this->otherlifeskill) {
                 $this->saveOtherLifeSkill($this->otherlifeskill);
@@ -218,7 +211,36 @@ class CapasController extends Controller
                 $model->setAttribute('lifeskill', implode(', ', $data_skill));
             }
 
-            if ($model->save()) {
+            //photo
+            if (($model->front_photo != $param['front_photo']) && $param['front_photo'] != null) {
+                //delte old photo
+                $this->photo->deleteProntPhoto($model->front_photo);
+                $this->photo->deleteProntPhotoThumb($model->front_photo);
+
+                //save new photo
+                $this->photo->copyProntPhoto($param['front_photo']);
+                $this->photo->copyProntPhotoThumb($param['front_photo']);
+            }
+
+            if (($model->side_photo != $param['side_photo']) && $param['side_photo'] != null) {
+                $this->photo->deleteSidePhoto($model->side_photo);
+                $this->photo->copySidePhoto($param['side_photo']);
+            }
+
+            if (($model->identity_card != $param['identity_card']) && $param['identity_card'] != null) {
+                $this->photo->deleteIdentityCardPhoto($model->identity_card);
+                $this->photo->copyIdentityCardPhoto($param['identity_card']);
+            }
+
+            if (($model->certificate_of_organization != $param['certificate_of_organization']) && $param['certificate_of_organization'] != null) {
+                $this->photo->deleteCertificatePhoto($model->certificate_of_organization);
+                $this->photo->copyCertificatePhoto($param['certificate_of_organization']);
+            }
+
+
+
+            //save database
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
 
                 if (null != $this->lifeskill) {
                     $model->saveTaxRelation($this->lifeskill, $model->id);
@@ -241,24 +263,27 @@ class CapasController extends Controller
                     $model->deleteAllTaxRelationByMemberId($data, $model->id);
                 }
 
-                return $this->redirect(['view', 'action' => 'member-capas-view', 'id' => $model->id]);
+                return $this->redirect(['view', 'action' => 'member-ppi-view', 'id' => $model->id]);
             } else {
                 return $this->render('update', [
-                    'model' => $model,
+                            'model' => $model
                 ]);
             }
         } else {
-            return $this->render('update', [
-                'model' => $model
-            ]);
+            throw new HttpException(403, 'You are not allowed to access this page', 0);
         }
     }
+
     public function actionTrash($id)
     {
-        $model = $this->findModel($id);
-        $model->save_status = 'Trash';
-        $model->save();
-        return $this->redirect(['index', 'action' => 'member-capas-list']);
+        if (Yii::$app->user->can('capastrash')) {
+            $model = $this->findModel($id);
+            $model->save_status = 'Trash';
+            $model->save();
+            return $this->redirect(['index', 'action' => 'member-capas-list']);
+        } else {
+            throw new HttpException(403, 'You are not allowed to access this page', 0);
+        }
     }
 
     /**
@@ -269,9 +294,13 @@ class CapasController extends Controller
      */
     public function actionDelete($id)
     {
-        $model = $this->findModel($id);
-        $model->delete();
-        return $this->redirect(['index', 'action' => 'member-capas-list']);
+        if (Yii::$app->user->can('capasdelete')) {
+            $model = $this->findModel($id);
+            $model->delete();
+            return $this->redirect(['index', 'action' => 'member-capas-list']);
+        } else {
+            throw new HttpException(403, 'You are not allowed to access this page', 0);
+        }
     }
 
     protected function trashAll($data)
@@ -322,79 +351,6 @@ class CapasController extends Controller
         }
     }
 
-    protected function copyProntPhoto($name)
-    {
-        if (null != $name) {
-//            return \yii\imagine\Image::thumbnail($this->original . $name, 215, 300)
-//            ->save($this->frontphoto . $name , ['quality' => 100]);
-            $img = new Image();
-            return $img->adaptiveResize($this->original . $name, $this->frontphoto . $name, 215, 300);
-        }
-    }
-    
-    protected function copyProntPhotoThumb($name)
-    {
-        if (null != $name) {
-//            return \yii\imagine\Image::thumbnail($this->original . $name, 215, 300)
-//            ->save($this->frontphoto . $name , ['quality' => 100]);
-            $img = new Image();
-            return $img->adaptiveResize($this->original . $name, $this->frontphoto . '42x42'. DS . $name, 42, 42);
-        }
-    }
-
-    protected function copySidePhoto($name)
-    {
-        if (null != $name) {
-//            return \yii\imagine\Image::thumbnail($this->original . $name, 215, 300)
-//                ->save($this->sidephoto . $name, ['quality' => 100]);
-            $img = new Image();
-            return $img->adaptiveResize($this->original . $name, $this->sidephoto . $name, 215, 300);
-        }
-    }
-
-    protected function copyCertificatePhoto($name)
-    {
-        if (null != $name) {
-            return copy($this->original . $name, $this->certificate . $name);
-        }
-    }
-
-    protected function copyIdentityCardPhoto($name)
-    {
-        if (null != $name) {
-            return copy($this->original . $name, $this->identitycard . $name);
-        }
-    }
-
-
-    protected function deleteProntPhoto($name)
-    {
-        if (null != $name) {
-            return unlink($this->frontphoto . $name);
-        }
-    }
-
-    protected function deleteSidePhoto($name)
-    {
-        if (null != $name) {
-            return unlink($this->sidephoto . $name);
-        }
-    }
-
-    protected function deleteCertificatePhoto($name)
-    {
-        if (null != $name) {
-            return unlink($this->certificate . $name);
-        }
-    }
-
-    protected function deleteIdentityCardPhoto($name)
-    {
-        if (null != $name) {
-            return unlink($this->identitycard . $name);
-        }
-    }
-
     protected function saveOtherLifeSkill($otherlifeskill)
     {
         $life = new LifeSkillModel;
@@ -406,4 +362,5 @@ class CapasController extends Controller
         $life->save();
         $this->otherlifeskillid = $life->id;
     }
+
 }
